Moved move struct to move.h and moved move.* to moveUtil.*
