+++ ChessAI/move/moveGeneration.h
@@ -2,28 +2,21 @@
 
 #include <vector>
 
+#include "../chess.h"
 #include "../enum.h"
-#include "../util/position.h"
 
 class ChessState;
 
 namespace move
 {
-	struct Move
-	{
-		util::Position source;
-		util::Position destination;
-		PieceType promotion;
+	struct Move;
 
-		Move() = default;
-		Move(const util::Position& source, const util::Position& destination);
-		Move(const util::Position& source, const util::Position& destination, const PieceType promotion);
-		Move(const Move& source);
-
-		Move& operator=(const Move& rightOperand);
-		bool operator==(const Move& rightOperand) const;
-		bool operator!=(const Move& rightOperand) const;
-	};
-
-	std::vector<Move> generatePawnMoves(const ChessState& game, const Color player);
+	/**
+	 * Generate moves for all of the specified player's pawns.
+	 *
+	 * \param chessState game state
+	 * \param player the player whose moves are being generated
+	 * \return valid moves for all the players pawns
+	 */
+	std::vector<Move> generatePawnMoves(const ChessState& chessState, const Color player);
 }
