+++ MoveTest/makeMoveTest.cpp
@@ -28,7 +28,7 @@ namespace makeMoveTest
 		const Position DESTINATION = SOURCE + RIGHT + UP * 2;
 		chessState.setState("4k3/8/8/8/8/8/4N3/4K3 w - - 0 1");
 
-		makeMove(COLOR, PIECE, DESTINATION, chessState);
+		makeMove(COLOR, SOURCE, DESTINATION, chessState);
 
 		EXPECT_FALSE(chessState.m_board.posIsOccupied(SOURCE));
 		EXPECT_TRUE(chessState.m_board.posIsOccupied(DESTINATION, COLOR, PIECE.m_pieceType));
@@ -43,7 +43,7 @@ namespace makeMoveTest
 		chessState.setState("4k3/8/8/8/8/8/4N3/4K3 w - - 0 1");
 		const Color CURRENT_TURN = chessState.getNextTurn();
 
-		makeMove(COLOR, PIECE, DESTINATION, chessState);
+		makeMove(COLOR, SOURCE, DESTINATION, chessState);
 
 		EXPECT_EQ(~CURRENT_TURN, chessState.getNextTurn());
 	}
@@ -56,7 +56,7 @@ namespace makeMoveTest
 		const Position DESTINATION = SOURCE + RIGHT + UP * 2;
 		chessState.setState("4k3/8/8/8/8/8/4N3/4K3 w - - 0 1");
 
-		makeMove(COLOR, PIECE, DESTINATION, chessState);
+		makeMove(COLOR, SOURCE, DESTINATION, chessState);
 
 		EXPECT_THAT(chessState.getWhitePieces(), Not(Contains(PIECE)));
 		EXPECT_THAT(chessState.getWhitePieces(), Contains(PieceNode(DESTINATION, PIECE.m_pieceType)));
@@ -71,7 +71,7 @@ namespace makeMoveTest
 		chessState.setState("4k3/8/8/8/8/8/4NN3/4K3 w - - 0 1");
 		const int CURRENT_HALF_TURNS = chessState.m_numHalfTurns;
 
-		makeMove(COLOR, PIECE, DESTINATION, chessState);
+		makeMove(COLOR, SOURCE, DESTINATION, chessState);
 
 		EXPECT_EQ(CURRENT_HALF_TURNS + 1, chessState.m_numHalfTurns);
 	}
@@ -85,7 +85,7 @@ namespace makeMoveTest
 		chessState.setState("4k3/8/8/8/8/8/4n3/4K3 b - - 0 1");
 		const int CURRENT_FULL_TURNS = chessState.m_numFullTurns;
 
-		makeMove(COLOR, PIECE, DESTINATION, chessState);
+		makeMove(COLOR, SOURCE, DESTINATION, chessState);
 
 		EXPECT_EQ(CURRENT_FULL_TURNS + 1, chessState.m_numFullTurns);
 	}
@@ -99,7 +99,7 @@ namespace makeMoveTest
 		chessState.setState("4k3/8/8/8/8/8/4NN3/4K3 w - - 0 1");
 		chessState.m_moveHistory.resize(MAX_MOVE_HISTORY_SIZE, MoveHistoryNode());
 
-		makeMove(COLOR, PIECE, DESTINATION, chessState);
+		makeMove(COLOR, SOURCE, DESTINATION, chessState);
 
 		EXPECT_EQ(MAX_MOVE_HISTORY_SIZE, chessState.m_moveHistory.size());
 		EXPECT_EQ(MoveHistoryNode(COLOR, PIECE, DESTINATION), chessState.m_moveHistory.back());
@@ -122,7 +122,7 @@ namespace makeMoveTest
 		}
 		chessState.m_moveHistory[(MAX_MOVE_HISTORY_SIZE / 2) - 1] = MoveHistoryNode(COLOR, PIECE, DESTINATION);
 
-		makeMove(COLOR, PIECE, DESTINATION, chessState);
+		makeMove(COLOR, SOURCE, DESTINATION, chessState);
 
 		EXPECT_EQ(Color::NEUTRAL, chessState.getNextTurn());
 	}
@@ -144,7 +144,7 @@ namespace makeMoveTest
 		}
 		chessState.m_moveHistory[(MAX_MOVE_HISTORY_SIZE / 2) - 1] = MoveHistoryNode(COLOR, PIECE, DESTINATION + UP);
 
-		makeMove(COLOR, PIECE, DESTINATION, chessState);
+		makeMove(COLOR, SOURCE, DESTINATION, chessState);
 
 		EXPECT_NE(Color::NEUTRAL, chessState.getNextTurn());
 	}
@@ -157,7 +157,7 @@ namespace makeMoveTest
 		const Position DESTINATION = SOURCE + RIGHT + UP * 2;
 		chessState.setState("4k3/8/8/8/8/8/8/4K3 w - - 0 1");
 
-		EXPECT_THROW(makeMove(COLOR, PIECE, DESTINATION, chessState), std::exception);
+		EXPECT_THROW(makeMove(COLOR, SOURCE, DESTINATION, chessState), std::exception);
 	}
 
 	namespace capture
@@ -171,7 +171,7 @@ namespace makeMoveTest
 			chessState.setState("4k3/8/8/8/5p2/8/4N3/4K3 w - - 0 1");
 
 			EXPECT_TRUE(chessState.m_board.posIsOccupied(DESTINATION, ~COLOR));
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(DESTINATION, ~COLOR));
 		}
 
@@ -184,7 +184,7 @@ namespace makeMoveTest
 			const PieceNode ENEMY = PieceNode(DESTINATION, PieceType::PAWN);
 			chessState.setState("4k3/8/8/8/5p2/8/4N3/4K3 w - - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_THAT(chessState.getBlackPieces(), Not(Contains(ENEMY)));
 		}
 
@@ -196,7 +196,7 @@ namespace makeMoveTest
 			const Position DESTINATION = SOURCE + RIGHT + UP * 2;
 			chessState.setState("4k3/8/8/8/5p2/8/4N3/4K3 w - - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_EQ(0, chessState.m_numHalfTurns);
 		}
 
@@ -209,7 +209,7 @@ namespace makeMoveTest
 			chessState.setState("r3k3/8/2N5/8/8/8/4N3/4K3 b q - 0 1");
 
 			EXPECT_TRUE(chessState.m_bQueenSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_bQueenSideCastle);
 		}
 
@@ -222,7 +222,7 @@ namespace makeMoveTest
 			chessState.setState("4k3/8/8/8/8/1n6/8/R3K3 w Q - 0 1");
 
 			EXPECT_TRUE(chessState.m_wQueenSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_wQueenSideCastle);
 		}
 
@@ -235,7 +235,7 @@ namespace makeMoveTest
 			chessState.setState("4k2r/8/6N1/8/8/8/8/4K3 b k - 0 1");
 
 			EXPECT_TRUE(chessState.m_bKingSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_bKingSideCastle);
 		}
 
@@ -248,7 +248,7 @@ namespace makeMoveTest
 			chessState.setState("4k3/8/8/8/8/6n1/8/4K2R w K - 0 1");
 
 			EXPECT_TRUE(chessState.m_wKingSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_wKingSideCastle);
 		}
 	}
@@ -263,7 +263,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(4, 5);
 			chessState.setState("4k3/8/8/8/8/8/4P3/4K3 w - - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_EQ(0, chessState.m_numHalfTurns);
 		}
 
@@ -275,7 +275,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(4, 2);
 			chessState.setState("4k3/8/8/3Pp3/8/8/8/4K3 w - d6 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(4, 3, ~COLOR));
 		}
 
@@ -287,7 +287,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(4, 5);
 			chessState.setState("4k3/8/8/8/3pP3/8/8/4K3 b - d3 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(4, 4, ~COLOR));
 		}
 
@@ -299,7 +299,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(4, 2);
 			chessState.setState("4k3/8/8/3Pp3/8/8/8/4K3 w - d6 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_THAT(chessState.getBlackPieces(), Not(Contains(PieceNode(DESTINATION + UP, PieceType::PAWN))));
 		}
 
@@ -311,7 +311,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(4, 5);
 			chessState.setState("4k3/8/8/8/3pP3/8/8/4K3 b - d3 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_THAT(chessState.getWhitePieces(), Not(Contains(PieceNode(DESTINATION + DOWN, PieceType::PAWN))));
 		}
 
@@ -324,7 +324,7 @@ namespace makeMoveTest
 			const PieceType PROMOTION = PieceType::KNIGHT;
 			chessState.setState("4k3/7P/8/8/8/8/8/4K3 w - - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState, PROMOTION);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState, PROMOTION);
 
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(SOURCE));
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(DESTINATION, COLOR, PieceType::PAWN));
@@ -339,7 +339,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(7, 0);
 			chessState.setState("4k3/7P/8/8/8/8/8/4K3 w - - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState, PieceType::KNIGHT);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState, PieceType::KNIGHT);
 			EXPECT_THAT(chessState.getWhitePieces(), Not(Contains(PieceNode(DESTINATION + UP, PieceType::PAWN))));
 		}
 
@@ -352,7 +352,7 @@ namespace makeMoveTest
 			const PieceType PROMOTION = PieceType::QUEEN;
 			chessState.setState("4k3/8/8/8/8/8/7p/4K3 b - - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(SOURCE));
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(DESTINATION, COLOR, PieceType::PAWN));
@@ -368,7 +368,7 @@ namespace makeMoveTest
 			const PieceType PROMOTION = PieceType::QUEEN;
 			chessState.setState("4k3/8/8/8/8/8/7p/4K3 b - - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_THAT(chessState.getBlackPieces(), Not(Contains(PieceNode(DESTINATION + DOWN, PieceType::PAWN))));
 		}
 	}
@@ -385,7 +385,7 @@ namespace makeMoveTest
 
 			EXPECT_TRUE(chessState.m_wKingSideCastle);
 			EXPECT_TRUE(chessState.m_wQueenSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState, PieceType::KNIGHT);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState, PieceType::KNIGHT);
 			EXPECT_FALSE(chessState.m_wKingSideCastle);
 			EXPECT_FALSE(chessState.m_wQueenSideCastle);
 		}
@@ -400,7 +400,7 @@ namespace makeMoveTest
 
 			EXPECT_TRUE(chessState.m_bKingSideCastle);
 			EXPECT_TRUE(chessState.m_bQueenSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_bKingSideCastle);
 			EXPECT_FALSE(chessState.m_bQueenSideCastle);
 		}
@@ -413,7 +413,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(2, 7);
 			chessState.setState("4k3/7P/8/8/8/8/8/R3K2R w KQ - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(0, 7));
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(SOURCE));
 			EXPECT_TRUE(chessState.m_board.posIsOccupied(DESTINATION, COLOR, PieceType::KING));
@@ -428,7 +428,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(2, 7);
 			chessState.setState("4k3/7P/8/8/8/8/8/R3K2R w KQ - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_THAT(chessState.getWhitePieces(), Contains(PieceNode(DESTINATION, PieceType::KING)));
 			EXPECT_THAT(chessState.getWhitePieces(), Contains(PieceNode(DESTINATION + RIGHT, PieceType::ROOK)));
 		}
@@ -441,7 +441,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(6, 7);
 			chessState.setState("4k3/7P/8/8/8/8/8/R3K2R w KQ - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(7, 7));
 			EXPECT_FALSE(chessState.m_board.posIsOccupied(SOURCE));
 			EXPECT_TRUE(chessState.m_board.posIsOccupied(DESTINATION, COLOR, PieceType::KING));
@@ -456,7 +456,7 @@ namespace makeMoveTest
 			const Position DESTINATION = Position(6, 7);
 			chessState.setState("4k3/7P/8/8/8/8/8/R3K2R w KQ - 0 1");
 
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_THAT(chessState.getWhitePieces(), Contains(PieceNode(DESTINATION, PieceType::KING)));
 			EXPECT_THAT(chessState.getWhitePieces(), Contains(PieceNode(DESTINATION + LEFT, PieceType::ROOK)));
 		}
@@ -473,7 +473,7 @@ namespace makeMoveTest
 			chessState.setState("4k3/8/8/8/8/8/8/R3K3 w Q - 0 1");
 
 			EXPECT_TRUE(chessState.m_wQueenSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_wQueenSideCastle);
 		}
 
@@ -486,7 +486,7 @@ namespace makeMoveTest
 			chessState.setState("r3k3/8/8/8/8/8/8/4K3 b q - 0 1");
 
 			EXPECT_TRUE(chessState.m_bQueenSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_bQueenSideCastle);
 		}
 
@@ -499,7 +499,7 @@ namespace makeMoveTest
 			chessState.setState("4k3/8/8/8/8/8/8/4K2R w K - 0 1");
 
 			EXPECT_TRUE(chessState.m_wKingSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_wKingSideCastle);
 		}
 
@@ -512,7 +512,7 @@ namespace makeMoveTest
 			chessState.setState("4k2r/8/8/8/8/8/8/4K3 b k - 0 1");
 
 			EXPECT_TRUE(chessState.m_bKingSideCastle);
-			makeMove(COLOR, PIECE, DESTINATION, chessState);
+			makeMove(COLOR, SOURCE, DESTINATION, chessState);
 			EXPECT_FALSE(chessState.m_bKingSideCastle);
 		}
 	}
